Construct a Paillier public and private key pair given a fixed number of bits

Construct a Paillier public and private key pair given a fixed number of bits

Construct a Paillier public and private key pair given a fixed number of bits

Format

An R6Class generator object

Methods

PaillierKeyPair$getPrivateKey()

Return the private key

See also

Public fields

pubkey

the public key

Methods

Public methods

PaillierKeyPair$new()

PaillierKeyPair$getPrivateKey()

PaillierKeyPair$clone()

Method new()

Create a new public private key pair with specified number of modulus bits

Usage

PaillierKeyPair$new(modulusBits)

Arguments

modulusBits

the number of bits to use

Returns

a PaillierKeyPair object

Method getPrivateKey()

Return the private key

Usage

PaillierKeyPair$getPrivateKey()

Returns

the private key

Method clone()

The objects of this class are cloneable with this method.

Usage

PaillierKeyPair$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

keys <- PaillierKeyPair$new(1024)
keys$pubkey
#> <PaillierPublicKey>
#>   Public:
#>     add: function (a, b) 
#>     add_real: function (den, a, b) 
#>     bits: 1024
#>     clone: function (deep = FALSE) 
#>     encrypt: function (m) 
#>     initialize: function (bits, n) 
#>     mult: function (a, b) 
#>     n: 10184537563970004142560433767615850729771478490462661106 ...
#>     nPlusOne: 10184537563970004142560433767615850729771478490462661106 ...
#>     nSquared: 10372480539191606622228559744429889173923933753767522965 ...
#>     sub: function (a, b) 
#>     sub_real: function (den, a, b) 
#>   Private:
#>     randomize: function (a) 
keys$getPrivateKey()
#> <PaillierPrivateKey>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     decrypt: function (c) 
#>     getLambda: function () 
#>     initialize: function (lambda, pubkey) 
#>     pubkey: PaillierPublicKey, R6
#>   Private:
#>     lambda: 50922687819850020712802168838079253648857392452313305534 ...
#>     x: 98038833236142070585609372374786714275167968114207583133 ...