Construct a Paillier public and private key pair given a fixed number of bits

Construct a Paillier public and private key pair given a fixed number of bits

Construct a Paillier public and private key pair given a fixed number of bits

Format

An R6::R6Class() generator object

Methods

PaillierKeyPair$getPrivateKey()

Return the private key

See also

Public fields

pubkey

the public key

Methods

Public methods

PaillierKeyPair$new()

PaillierKeyPair$getPrivateKey()

PaillierKeyPair$clone()

Method new()

Create a new public private key pair with specified number of modulus bits

Usage

PaillierKeyPair$new(modulusBits)

Arguments

modulusBits

the number of bits to use

Returns

a PaillierKeyPair object

Method getPrivateKey()

Return the private key

Usage

PaillierKeyPair$getPrivateKey()

Returns

the private key

Method clone()

The objects of this class are cloneable with this method.

Usage

PaillierKeyPair$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

keys <- PaillierKeyPair$new(1024)
keys$pubkey
#> <PaillierPublicKey>
#>   Public:
#>     add: function (a, b) 
#>     add_real: function (den, a, b) 
#>     bits: 1024
#>     clone: function (deep = FALSE) 
#>     encrypt: function (m) 
#>     initialize: function (bits, n) 
#>     mult: function (a, b) 
#>     n: 11820012908153722257005420763105883377048387354998106036 ...
#>     nPlusOne: 11820012908153722257005420763105883377048387354998106036 ...
#>     nSquared: 13971270514892061458812556422406434400778776578339315363 ...
#>     sub: function (a, b) 
#>     sub_real: function (den, a, b) 
#>   Private:
#>     randomize: function (a) 
keys$getPrivateKey()
#> <PaillierPrivateKey>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     decrypt: function (c) 
#>     getLambda: function () 
#>     initialize: function (lambda, pubkey) 
#>     pubkey: PaillierPublicKey, R6
#>   Private:
#>     lambda: 19700021513589537095009034605176472295080645591663510060 ...
#>     x: 69190289511327758037920180744386648793875527320026815702 ...